main.py
import os
from fastapi import FastAPI, HTTPException, Request
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import uvicorn
from dotenv import load_dotenv
import requests

load_dotenv()

app = FastAPI(title="Visionary API")

# Enable CORS for frontend integration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

class SearchQuery(BaseModel):
    query: str

class PropertyResult(BaseModel):
    id: int
    title: str
    location: str
    price: str
    meta: str
    beforeImage: str
    afterImage: str

@app.get("/")
async def root():
    return {"message": "Visionary Core API Active"}

@app.post("/api/search", response_model=List[PropertyResult])
async def search_properties(query: SearchQuery):
    # In a real MVP, this would scrape or hit a real estate API.
    # For now, we simulate the 'Grounding' process.
    print(f"Searching for: {query.query}")
    
    # Mock results based on real patterns
    return [
        {
            "id": 101,
            "title": "Heritage Restoration",
            "location": f"{query.query}",
            "price": "$1.1M - $1.25M",
            "meta": "3 BEDS | 2 BATH | HERITAGE",
            "beforeImage": "terrace_house_before_1768693729936.png",
            "afterImage": "terrace_house_after_v2_1768693859721.png"
        }
    ]

@app.post("/api/vision/generate")
async def generate_vision(property_id: int, style: str = "Architectural"):
    # This endpoint would trigger the Nano Banana (Gemini) edit process
    # using the developer's background API keys.
    return {
        "status": "success",
        "message": f"Vision in style '{style}' generated for property {property_id}.",
        "integrity_score": 0.99
    }

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
